requirements.txt
fire==0.5.0
lmdb==1.4.1
matplotlib==3.7.1
natsort==8.4.0
nltk==3.8.1
numpy==1.23.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.8.1.78
pandas==1.5.3
Pillow==9.3.0
Pillow==10.1.0
six==1.16.0
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.220
warpctc_pytorch==0.2.1
